+++ headers/bios/io.h
@@ -4,7 +4,7 @@
 #include "headers/types.h"
 #include "headers/support/isoreader.h" //Need for structure!
 
-typedef int (*SECTORHANDLER)(char *filename,uint_32 sector, void *buffer); //Write/read a 512-byte sector! Result=1 on success, 0 on error!
+typedef byte (*SECTORHANDLER)(char *filename,uint_32 sector, void *buffer); //Write/read a 512-byte sector! Result=1 on success, 0 on error!
 typedef void(*DISKCHANGEDHANDLER)(int disk); //Disk has been changed!
 
 typedef struct
@@ -37,10 +37,10 @@ void iohdd0(char *filename, uint_64 startpos, byte readonly, uint_32 customsize)
 void iohdd1(char *filename, uint_64 startpos, byte readonly, uint_32 customsize);
 void iocdrom0(char *filename, uint_64 startpos, byte readonly, uint_32 customsize);
 void iocdrom1(char *filename, uint_64 startpos, byte readonly, uint_32 customsize);
-int readdata(int device, void *buffer, uint_64 startpos, uint_32 bytestoread);
-int writedata(int device, void *buffer, uint_64 startpos, uint_32 bytestowrite);
-int has_drive(int drive); //Have drive?
-int drivereadonly(int drive); //Drive is read-only?
+byte readdata(int device, void *buffer, uint_64 startpos, uint_32 bytestoread);
+byte writedata(int device, void *buffer, uint_64 startpos, uint_32 bytestowrite);
+byte has_drive(int drive); //Have drive?
+byte drivereadonly(int drive); //Drive is read-only?
 FILEPOS getdisksize(int device); //Retrieve a dynamic/static image size!
 uint_64 disksize(int disknumber); //Currently mounted disk size!
 void register_DISKCHANGE(int device, DISKCHANGEDHANDLER diskchangedhandler); //Register a disk changed handler!
